+++ webrtc/modules/audio_coding/main/acm2/acm_receiver_unittest_oldapi.cc
@@ -124,7 +124,7 @@ class AcmReceiverTestOldApi : public AudioPacketizationCallback,
       uint8_t payload_type,
       uint32_t timestamp,
       const uint8_t* payload_data,
-      uint16_t payload_len_bytes,
+      size_t payload_len_bytes,
       const RTPFragmentationHeader* fragmentation) OVERRIDE {
     if (frame_type == kFrameEmpty)
       return 0;
