+++ webrtc/modules/audio_coding/main/acm2/acm_send_test.h
@@ -49,7 +49,7 @@ class AcmSendTest : public AudioPacketizationCallback, public PacketSource {
       uint8_t payload_type,
       uint32_t timestamp,
       const uint8_t* payload_data,
-      uint16_t payload_len_bytes,
+      size_t payload_len_bytes,
       const RTPFragmentationHeader* fragmentation) OVERRIDE;
 
  private:
