+++ webrtc/modules/audio_coding/main/acm2/acm_send_test_oldapi.cc
@@ -98,7 +98,7 @@ int32_t AcmSendTestOldApi::SendData(
     uint8_t payload_type,
     uint32_t timestamp,
     const uint8_t* payload_data,
-    uint16_t payload_len_bytes,
+    size_t payload_len_bytes,
     const RTPFragmentationHeader* fragmentation) {
   // Store the packet locally.
   frame_type_ = frame_type;
