+++ webrtc/modules/audio_coding/main/acm2/acm_send_test_oldapi.h
@@ -51,7 +51,7 @@ class AcmSendTestOldApi : public AudioPacketizationCallback,
       uint8_t payload_type,
       uint32_t timestamp,
       const uint8_t* payload_data,
-      uint16_t payload_len_bytes,
+      size_t payload_len_bytes,
       const RTPFragmentationHeader* fragmentation) OVERRIDE;
 
  private:
