+++ webrtc/modules/audio_coding/main/test/RTPFile.h
@@ -28,12 +28,12 @@ class RTPStream {
 
   virtual void Write(const uint8_t payloadType, const uint32_t timeStamp,
                      const int16_t seqNo, const uint8_t* payloadData,
-                     const uint16_t payloadSize, uint32_t frequency) = 0;
+                     const size_t payloadSize, uint32_t frequency) = 0;
 
   // Returns the packet's payload size. Zero should be treated as an
   // end-of-stream (in the case that EndOfFile() is true) or an error.
-  virtual uint16_t Read(WebRtcRTPHeader* rtpInfo, uint8_t* payloadData,
-                        uint16_t payloadSize, uint32_t* offset) = 0;
+  virtual size_t Read(WebRtcRTPHeader* rtpInfo, uint8_t* payloadData,
+                      size_t payloadSize, uint32_t* offset) = 0;
   virtual bool EndOfFile() const = 0;
 
  protected:
@@ -46,7 +46,7 @@ class RTPStream {
 class RTPPacket {
  public:
   RTPPacket(uint8_t payloadType, uint32_t timeStamp, int16_t seqNo,
-            const uint8_t* payloadData, uint16_t payloadSize,
+            const uint8_t* payloadData, size_t payloadSize,
             uint32_t frequency);
 
   ~RTPPacket();
@@ -55,7 +55,7 @@ class RTPPacket {
   uint32_t timeStamp;
   int16_t seqNo;
   uint8_t* payloadData;
-  uint16_t payloadSize;
+  size_t payloadSize;
   uint32_t frequency;
 };
 
@@ -67,10 +67,10 @@ class RTPBuffer : public RTPStream {
 
   virtual void Write(const uint8_t payloadType, const uint32_t timeStamp,
                      const int16_t seqNo, const uint8_t* payloadData,
-                     const uint16_t payloadSize, uint32_t frequency) OVERRIDE;
+                     const size_t payloadSize, uint32_t frequency) OVERRIDE;
 
-  virtual uint16_t Read(WebRtcRTPHeader* rtpInfo, uint8_t* payloadData,
-                        uint16_t payloadSize, uint32_t* offset) OVERRIDE;
+  virtual size_t Read(WebRtcRTPHeader* rtpInfo, uint8_t* payloadData,
+                      size_t payloadSize, uint32_t* offset) OVERRIDE;
 
   virtual bool EndOfFile() const OVERRIDE;
 
@@ -99,10 +99,10 @@ class RTPFile : public RTPStream {
 
   virtual void Write(const uint8_t payloadType, const uint32_t timeStamp,
                      const int16_t seqNo, const uint8_t* payloadData,
-                     const uint16_t payloadSize, uint32_t frequency) OVERRIDE;
+                     const size_t payloadSize, uint32_t frequency) OVERRIDE;
 
-  virtual uint16_t Read(WebRtcRTPHeader* rtpInfo, uint8_t* payloadData,
-                        uint16_t payloadSize, uint32_t* offset) OVERRIDE;
+  virtual size_t Read(WebRtcRTPHeader* rtpInfo, uint8_t* payloadData,
+                      size_t payloadSize, uint32_t* offset) OVERRIDE;
 
   virtual bool EndOfFile() const OVERRIDE {
     return _rtpEOF;
